Refactor user class into entity
